-DEV WORK:  -Mac Compatibility fixes.  -Re-included GLEW  -Updated GLEW and libs to 1.5.4  -removed unneeded .lib files  -fixed some compiler warnings.  -updated various elements of the vcproj file.

+++ src/modelexport_obj.cpp
@@ -44,7 +44,7 @@ void ExportM2toOBJ(Attachment *att, Model *m, const char *fn, bool init)
 		ModelRenderPass &p = m->passes[i];
 			
 		if (p.init(m)) {
-			wxString texName = GetM2TextureName(m,fn,p,i);
+			wxString texName = GetM2TextureName(m,fn,p,(int)i);
 
 			fm << _T("newmtl ") << texName << endl;
 			texName << _T(".tga");
